+++ packages/react/rollup.config.ts
@@ -4,7 +4,7 @@ import typescript from '@rollup/plugin-typescript';
 import externals from 'rollup-plugin-node-externals';
 
 // common config settings
-const input = ['src/index.ts', 'src/internal.ts', 'src/theme.ts'];
+const input = ['src/index.ts', 'src/internal.ts', 'src/server.ts'];
 const sourceMap = false;
 const tsconfig = 'tsconfig.dist.json';
