+++ packages/ui/src/theme/createTheme/createTheme.ts
@@ -4,7 +4,7 @@ import flattenProperties from 'style-dictionary/lib/utils/flattenProperties.js';
 
 import { defaultTheme } from '../defaultTheme';
 import { Theme, DefaultTheme, WebTheme, Override } from '../types';
-import { setupToken, setupTokens } from './setupToken';
+import { setupToken, setupTokens } from './utils';
 import { WebDesignToken } from '../tokens/types/designToken';
 import { createComponentCSS } from './createComponentCSS';
 import { isString } from '../../utils';
