--- packages/ui/src/theme/createTheme/isShadowObject.ts
@@ -1,8 +1,0 @@
-import { ShadowValue } from '../tokens/types/designToken';
-import { isObject, has } from '../../utils';
-
-export function isShadowTokenObject(
-  value: unknown
-): value is ShadowValue & object {
-  return isObject(value) && has(value, 'offsetX');
-}
